+++ internal/resourceconsumption/host_cpu_stub.go
@@ -1,0 +1,7 @@
+//go:build !linux
+
+package resourceconsumption
+
+func (rcm *Manager) getTotalCPULinux() float64 {
+	return 0
+}